Refactor `RoutingSettings`.
