+++ src/request_parser.h
@@ -9,7 +9,7 @@
 #include "request_types.h"
 
 namespace rm {
-std::optional<RoutingSettings> ParseSettings(json::Dict settings);
+std::optional<RoutingSettings> ParseRoutingSettings(json::Dict settings);
 
 std::optional<std::vector<PostRequest>> ParseInput(json::List base_requests);
 std::optional<std::vector<GetRequest>> ParseOutput(json::List stat_requests);
